requirements.txt
langchain
openai
huggingface_hub
python-dotenv
langchain_openai
langchain-community
streamlit

